requirements.txt
django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
pytils==0.4.1
pytz==2024.1
python-dotenv==1.0.1
PyJWT==2.8.0